[IMP] base_exception: One model per python file
